fixed incorrect interpretation of wes' max edge cuts to be the right thing
